pyproject.toml
[tool.poetry]
name = "blockutils"
version = "0.1.0"
description = ""
authors = ["yen <yen@xy.finance>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bip-utils = "^2.9.3"
eth-account = "^0.13.0"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
